+++ Libplanet.Extensions.Cocona/Utils.cs
@@ -9,6 +9,7 @@ using System.Text.Json.Serialization;
 using Bencodex.Types;
 using global::Cocona;
 using Libplanet.Action;
+using Libplanet.Store;
 
 namespace Libplanet.Extensions.Cocona
 {
@@ -77,6 +78,60 @@ namespace Libplanet.Extensions.Cocona
                     row.Item2,
                     row.Item3,
                 }));
+        }
+
+        public static IStore LoadStoreFromUri(string uriString)
+        {
+            // TODO: Cocona supports .NET's TypeConverter protocol for instantiating objects
+            // from CLI options/arguments.  We'd better to implement it for IStore, and simply
+            // use IStore as the option/argument types rather than taking them as strings.
+            var uri = new Uri(uriString);
+
+            switch (uri.Scheme)
+            {
+                case "default":
+                case "rocksdb":
+                    Console.Error.WriteLine(
+                        "Warning: The scheme {0}:// for the store URI is deprecated.  " +
+                        "Use {0}+file:// instead.",
+                        uri.Scheme
+                    );
+                    Console.Error.Flush();
+                    uri = new Uri($"{uri.Scheme}+file{uri.ToString().Substring(7)}");
+                    break;
+
+                case "":
+                case null:
+                    const string? exceptionMessage = "A store URI must have a scheme, " +
+                                                     "e.g., default+file://, rocksdb+file://.";
+                    throw new ArgumentException(exceptionMessage, nameof(uriString));
+            }
+
+            if (StoreLoaderAttribute.LoadStore(uri) is { } pair)
+            {
+                pair.StateStore.Dispose();
+                return pair.Store;
+            }
+
+            (string UriScheme, Type DeclaringType)[] loaders =
+                StoreLoaderAttribute.ListStoreLoaders().ToArray();
+            int longestSchemeLength = Math.Max(
+                "URI SCHEME".Length,
+                loaders.Max(loader => loader.UriScheme.Length + 1)
+            );
+            string loaderTable = string.Join(
+                "\n",
+                loaders.Select(pair =>
+                    $"  {(pair.UriScheme + ":").PadRight(longestSchemeLength)}" +
+                    $"  {pair.DeclaringType.Name}")
+            );
+
+            throw new ArgumentException(
+                $"The store URI scheme {uri.Scheme}:// is not supported.  See also " +
+                $"the list of supported store URI schemes:\n\n" +
+                $"  {"URI SCHEME".PadRight(longestSchemeLength)}  PROVIDED BY\n{loaderTable}\n",
+                nameof(uriString)
+            );
         }
 
         public static string SerializeHumanReadable<T>(T target) =>